[DEL] viewmodel관련 필요없는 파일삭제(#162)
