--- LionHeart-iOS/LionHeart-iOS/Scenes/Login/LoginViewModelable.swift
@@ -1,20 +1,0 @@
-//
-//  LoginViewModelable.swift
-//  LionHeart-iOS
-//
-//  Created by 김민재 on 11/4/23.
-//
-
-import Foundation
-import Combine
-
-protocol LoginViewModel: ViewModel where Input == LoginViewModelInput, Output == LoginViewModelOutput {}
-
-struct LoginViewModelInput {
-    let kakakoLoginButtonTap: PassthroughSubject<Void, Never>
-}
-
-struct LoginViewModelOutput {
-    let loginSuccess: AnyPublisher<String, Never>
-}
-
